chore: Put the thunks on another file, #34
